+++ src/Screens/Intro/styles.js
