+++ src/Screens/Mapa/styles.js
